+++ cmd/config/config.go
@@ -20,6 +20,7 @@ import (
 	"github.com/rocajuanma/anvil/cmd/config/pull"
 	"github.com/rocajuanma/anvil/cmd/config/push"
 	"github.com/rocajuanma/anvil/cmd/config/show"
+	"github.com/rocajuanma/anvil/cmd/config/sync"
 	"github.com/rocajuanma/anvil/pkg/constants"
 	"github.com/spf13/cobra"
 )
@@ -34,8 +35,9 @@ var ConfigCmd = &cobra.Command{
 }
 
 func init() {
-	// Add pull, push, and show as sub-commands of config
+	// Add pull, push, show, and sync as sub-commands of config
 	ConfigCmd.AddCommand(pull.PullCmd)
 	ConfigCmd.AddCommand(push.PushCmd)
 	ConfigCmd.AddCommand(show.ShowCmd)
+	ConfigCmd.AddCommand(sync.SyncCmd)
 }